refactor(meta): extract utilities from monolithic util.rs
Transform the large util.rs file (561 lines) into 8 focused utility modules
to improve code organization, maintainability, and developer experience.

PROBLEM:
- Monolithic util.rs with 561 lines and 16 scattered functions
- Inconsistent import patterns across 39 API files
- 165+ occurrences of duplicated transaction utilities
- Poor discoverability and maintainability

SOLUTION:
Applied incremental refactoring with backward compatibility:

EXTRACTED MODULES:
- txn_condition_util.rs - Transaction condition builders (txn_cond_eq_seq, txn_cond_seq)
- txn_retry_util.rs - Transaction retry patterns (txn_backoff re-export)
- serialization_util.rs - Data serialization (serialize_struct, deserialize_struct, etc.)
- error_util.rs - Error construction (6 functions from util.rs, database_util.rs, schema_api.rs)
- kv_fetch_util.rs - KV operations (get_u64_value, fetch_id, 4 others)
- txn_op_builder_util.rs - Transaction builders (txn_put_pb, txn_op_del, 3 others)
- txn_core_util.rs - Core transactions (send_txn, txn_delete_exact, txn_replace_exact)
- data_retention_util.rs - Retention logic (is_drop_time_retainable, get_retention_boundary)
