+++ src/meta/api/src/database_api.rs
@@ -55,21 +55,21 @@ use log::error;
 use log::warn;
 use seq_marked::SeqValue;
 
-use crate::database_util::db_has_to_not_exist;
+use crate::data_retention_util::is_drop_time_retainable;
 use crate::database_util::drop_database_meta;
 use crate::database_util::get_db_or_err;
 use crate::db_has_to_exist;
+use crate::error_util::db_has_to_not_exist;
 use crate::fetch_id;
 use crate::kv_app_error::KVAppError;
 use crate::kv_pb_api::KVPbApi;
-use crate::send_txn;
 use crate::serialize_struct;
 use crate::serialize_u64;
 use crate::txn_backoff::txn_backoff;
-use crate::txn_cond_seq;
+use crate::txn_condition_util::txn_cond_seq;
+use crate::txn_core_util::send_txn;
 use crate::txn_op_del;
 use crate::txn_op_put;
-use crate::util::is_drop_time_retainable;
 
 impl<KV> DatabaseApi for KV
 where
