+++ src/meta/api/src/dictionary_api.rs
@@ -38,11 +38,11 @@ use crate::kv_app_error::KVAppError;
 use crate::kv_pb_api::KVPbApi;
 use crate::meta_txn_error::MetaTxnError;
 use crate::name_id_value_api::NameIdValueApi;
-use crate::send_txn;
 use crate::txn_backoff::txn_backoff;
-use crate::txn_cond_seq;
+use crate::txn_condition_util::txn_cond_seq;
+use crate::txn_core_util::send_txn;
+use crate::txn_op_builder_util::txn_op_put_pb;
 use crate::txn_op_del;
-use crate::util::txn_op_put_pb;
 
 /// DictionaryApi defines APIs for dictionary management.
 ///
