+++ src/meta/api/src/garbage_collection_api.rs
@@ -54,15 +54,15 @@ use log::info;
 use crate::index_api::IndexApi;
 use crate::kv_app_error::KVAppError;
 use crate::kv_pb_api::KVPbApi;
-use crate::send_txn;
 use crate::txn_backoff::txn_backoff;
-use crate::txn_cond_eq_seq;
+use crate::txn_condition_util::txn_cond_eq_seq;
+use crate::txn_core_util::send_txn;
+use crate::txn_core_util::txn_delete_exact;
+use crate::txn_core_util::txn_replace_exact;
+use crate::txn_op_builder_util::txn_op_put_pb;
+use crate::txn_op_builder_util::txn_put_pb;
 use crate::txn_op_del;
 use crate::txn_op_get;
-use crate::util::txn_delete_exact;
-use crate::util::txn_op_put_pb;
-use crate::util::txn_put_pb;
-use crate::util::txn_replace_exact;
 
 /// GarbageCollectionApi defines APIs for garbage collection operations.
 ///
