+++ src/meta/api/src/lock_api.rs
@@ -37,9 +37,9 @@ use crate::kv_app_error::KVAppError;
 use crate::kv_pb_api::KVPbApi;
 use crate::kv_pb_crud_api::KVPbCrudApi;
 use crate::txn_backoff::txn_backoff;
-use crate::txn_cond_seq;
+use crate::txn_condition_util::txn_cond_seq;
+use crate::txn_op_builder_util::txn_op_put_pb;
 use crate::txn_op_put;
-use crate::util::txn_op_put_pb;
 use crate::util::IdempotentKVTxnResponse;
 use crate::util::IdempotentKVTxnSender;
