+++ src/meta/api/src/name_id_value_api.rs
@@ -35,16 +35,16 @@ use fastrace::func_name;
 use futures::TryStreamExt;
 use log::debug;
 
+use crate::kv_fetch_util::fetch_id;
 use crate::kv_pb_api::KVPbApi;
 use crate::kv_pb_api::UpsertPB;
 use crate::meta_txn_error::MetaTxnError;
 use crate::txn_backoff::txn_backoff;
+use crate::txn_condition_util::txn_cond_eq_seq;
+use crate::txn_core_util::send_txn;
+use crate::txn_core_util::txn_delete_exact;
+use crate::txn_op_builder_util::txn_op_put_pb;
 use crate::txn_op_del;
-use crate::util::fetch_id;
-use crate::util::send_txn;
-use crate::util::txn_cond_eq_seq;
-use crate::util::txn_delete_exact;
-use crate::util::txn_op_put_pb;
 
 /// NameIdValueApi provide generic meta-service access pattern implementations for `name -> id -> value` mapping.
 ///
