+++ src/meta/api/src/row_access_policy_api_impl.rs
@@ -33,11 +33,11 @@ use crate::fetch_id;
 use crate::kv_pb_api::KVPbApi;
 use crate::meta_txn_error::MetaTxnError;
 use crate::row_access_policy_api::RowAccessPolicyApi;
-use crate::send_txn;
 use crate::txn_backoff::txn_backoff;
-use crate::txn_cond_eq_seq;
-use crate::util::txn_delete_exact;
-use crate::util::txn_op_put_pb;
+use crate::txn_condition_util::txn_cond_eq_seq;
+use crate::txn_core_util::send_txn;
+use crate::txn_core_util::txn_delete_exact;
+use crate::txn_op_builder_util::txn_op_put_pb;
 
 /// RowAccessPolicyApi is implemented upon kvapi::KVApi.
 /// Thus every type that impl kvapi::KVApi impls RowAccessPolicyApi.
