+++ src/meta/api/src/serialization_util.rs
@@ -1,0 +1,64 @@
+// Copyright 2021 Datafuse Labs
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+//! Serialization utilities extracted from util.rs
+
+use databend_common_meta_app::primitive::Id;
+use databend_common_meta_types::InvalidArgument;
+use databend_common_meta_types::InvalidReply;
+use databend_common_meta_types::MetaNetworkError;
+use databend_common_proto_conv::FromToProto;
+
+pub fn serialize_u64(value: impl Into<Id>) -> Result<Vec<u8>, MetaNetworkError> {
+    let v = serde_json::to_vec(&*value.into()).map_err(|e| {
+        let inv = InvalidArgument::new(e, "");
+        MetaNetworkError::InvalidArgument(inv)
+    })?;
+    Ok(v)
+}
+
+pub fn deserialize_u64(v: &[u8]) -> Result<Id, MetaNetworkError> {
+    let id = serde_json::from_slice(v).map_err(|e| {
+        let inv = InvalidReply::new("", &e);
+        MetaNetworkError::InvalidReply(inv)
+    })?;
+    Ok(Id::new(id))
+}
+
+pub fn serialize_struct<T>(value: &T) -> Result<Vec<u8>, MetaNetworkError>
+where T: FromToProto + 'static {
+    let p = value.to_pb().map_err(|e| {
+        let inv = InvalidArgument::new(e, "");
+        MetaNetworkError::InvalidArgument(inv)
+    })?;
+    let mut buf = vec![];
+    prost::Message::encode(&p, &mut buf).map_err(|e| {
+        let inv = InvalidArgument::new(e, "");
+        MetaNetworkError::InvalidArgument(inv)
+    })?;
+    Ok(buf)
+}
+
+pub fn deserialize_struct<T>(buf: &[u8]) -> Result<T, MetaNetworkError>
+where T: FromToProto {
+    let p: T::PB = prost::Message::decode(buf).map_err(|e| {
+        let inv = InvalidReply::new("", &e);
+        MetaNetworkError::InvalidReply(inv)
+    })?;
+    let v: T = FromToProto::from_pb(p).map_err(|e| {
+        let inv = InvalidReply::new("", &e);
+        MetaNetworkError::InvalidReply(inv)
+    })?;
+    Ok(v)
+}
