+++ src/meta/api/src/table_api.rs
@@ -113,11 +113,15 @@ use crate::assert_table_exist;
 use crate::database_api::DatabaseApi;
 use crate::database_util::get_db_or_err;
 use crate::deserialize_struct;
+use crate::error_util::table_has_to_not_exist;
 use crate::fetch_id;
 use crate::garbage_collection_api::get_history_tables_for_gc;
 use crate::garbage_collection_api::ORPHAN_POSTFIX;
 use crate::get_u64_value;
 use crate::kv_app_error::KVAppError;
+use crate::kv_fetch_util::deserialize_id_get_response;
+use crate::kv_fetch_util::deserialize_struct_get_response;
+use crate::kv_fetch_util::mget_pb_values;
 use crate::kv_pb_api::KVPbApi;
 use crate::kv_pb_crud_api::KVPbCrudApi;
 use crate::list_u64_value;
@@ -127,20 +131,16 @@ use crate::schema_api::get_db_by_id_or_err;
 use crate::schema_api::get_db_id_or_err;
 use crate::schema_api::get_history_table_metas;
 use crate::schema_api::handle_undrop_table;
-use crate::schema_api::table_has_to_not_exist;
-use crate::send_txn;
 use crate::serialize_struct;
 use crate::serialize_u64;
 use crate::txn_backoff::txn_backoff;
-use crate::txn_cond_eq_seq;
-use crate::txn_cond_seq;
+use crate::txn_condition_util::txn_cond_eq_seq;
+use crate::txn_condition_util::txn_cond_seq;
+use crate::txn_core_util::send_txn;
+use crate::txn_op_builder_util::txn_op_put_pb;
 use crate::txn_op_del;
 use crate::txn_op_get;
 use crate::txn_op_put;
-use crate::util::deserialize_id_get_response;
-use crate::util::deserialize_struct_get_response;
-use crate::util::mget_pb_values;
-use crate::util::txn_op_put_pb;
 use crate::util::IdempotentKVTxnResponse;
 use crate::util::IdempotentKVTxnSender;
 use crate::DEFAULT_MGET_SIZE;
